Restructure edit form file location
